+++ core/server/api_container/server/service_network/user_service_launcher/user_service_launcher_test.go
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) 2021 - present Kurtosis Technologies Inc.
- * All Rights Reserved.
- */
-
-package user_service_launcher