+++ core/docs/changelog.md
@@ -36,7 +36,7 @@
 
 ### Changes
 * Replaced  parameter `DockerManager` with `KurtosisBackend` in `standardSidecarExecCmdExecutor`
-* Upgraded to container-engine-lib xxxx, which brings the new `DockerKurtosisBackend` implementation to replace the `DockerManager` //TODO UPDATE VERSION HERE
+* Upgraded to container-engine-lib 0.14.1
 * Updated api_container_launcher to use kurtosis_backend for launching API container instances
 * Update user_service_launcher to use kurtosis_backend for launching services
 * Updated module_store and module_launcher to use kurtosis_backend for module operations